+++ src/test/kotlin/com/zepben/evolve/streaming/get/QueryNetworkStateClientTest.kt
@@ -8,9 +8,7 @@
 
 package com.zepben.evolve.streaming.get
 
-import com.zepben.evolve.streaming.data.CurrentStateEventBatch
-import com.zepben.evolve.streaming.data.SwitchAction
-import com.zepben.evolve.streaming.data.SwitchStateEvent
+import com.zepben.evolve.streaming.data.*
 import com.zepben.evolve.streaming.get.testservices.TestQueryNetworkStateService
 import com.zepben.evolve.streaming.grpc.GrpcChannel
 import com.zepben.evolve.streaming.grpc.TokenCallCredentials
@@ -18,11 +16,16 @@ import com.zepben.protobuf.ns.QueryNetworkStateServiceGrpc
 import io.grpc.inprocess.InProcessChannelBuilder
 import io.grpc.inprocess.InProcessServerBuilder
 import io.grpc.testing.GrpcCleanupRule
+import io.mockk.slot
+import io.mockk.spyk
+import io.mockk.verifySequence
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
+import org.hamcrest.Matchers.instanceOf
 import org.junit.Rule
 import org.junit.jupiter.api.Test
-import org.mockito.kotlin.spy
+import org.junit.jupiter.api.assertTimeoutPreemptively
+import java.time.Duration
 import java.time.LocalDateTime
 import kotlin.streams.toList
 
@@ -34,7 +37,7 @@ internal class QueryNetworkStateClientTest {
 
     private val serverName = InProcessServerBuilder.generateName()
     private val channel = grpcCleanup.register(InProcessChannelBuilder.forName(serverName).directExecutor().build())
-    private val stub = spy(QueryNetworkStateServiceGrpc.newStub(channel))
+    private val stub = spyk(QueryNetworkStateServiceGrpc.newStub(channel))
     private val client = QueryNetworkStateClient(stub, null)
     private val service = TestQueryNetworkStateService()
 
@@ -58,6 +61,30 @@ internal class QueryNetworkStateClientTest {
         testGetCurrentStates { from, to -> client.getCurrentStatesStream(1, from, to).toList() }
     }
 
+    @Test
+    internal fun `can report batch status`() {
+        val status = BatchSuccessful(1234)
+        service.onBatchStatus = spyk()
+
+        // To check the `onCompleted` is correctly waited we rely on the call to `reportBatchStatus` not timing out while it waits.
+        assertTimeoutPreemptively(
+            Duration.ofSeconds(5),
+            message = "If this test times out, the `onCompletes` loop hasn't triggered."
+        ) {
+            client.reportBatchStatus(status)
+        }
+
+        val received = slot<SetCurrentStatesStatus>()
+        verifySequence {
+            service.onBatchStatus(capture(received))
+        }
+
+        received.captured.also {
+            assertThat(it, instanceOf(BatchSuccessful::class.java))
+            assertThat(it.batchId, equalTo(status.batchId))
+        }
+    }
+
     private fun testGetCurrentStates(act: (LocalDateTime, LocalDateTime) -> List<CurrentStateEventBatch>) {
         val from = LocalDateTime.now().plusDays(-1)
         val to = LocalDateTime.now()
@@ -68,11 +95,12 @@ internal class QueryNetworkStateClientTest {
             CurrentStateEventBatch(4, listOf(SwitchStateEvent("event4", from.plusHours(1), "switch-2", SwitchAction.CLOSE)))
         )
 
-        service.onGetCurrentStates = spy { _, _ -> batches.asSequence() }
+        service.onGetCurrentStates = spyk({ _, _ -> batches.asSequence() })
 
         val result = act(from, to)
 
         assertThat(result, equalTo(batches))
+        verifySequence { service.onGetCurrentStates(from, to) }
     }
 
 }